Fixed bug where puting in a String for team name returns team number 0
A bug was discovered while testing the form in that when a person would
enter a string in for the 'Team Number', it would do an ajax lookup for
the team number of the provided string. However, the team lookup is done
via Laravel's Eloquent model system and therefore when looking up the
team's number (team_number int(11)) in the database, it would try and
convert the string into a number. The closest numeric representation of
the string 'Test' is the number 0, so it would return the team with a team
number of zero.